requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pyTelegramBotAPI==3.7.6
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0